refactor: 아키텍처 폴더 조정
