+++ src/widgets/feed-main-header/ui/FeedMainHeader.scss
