+++ src/widgets/feed/consts/type.ts
