+++ src/widgets/feed/ui/Feed.scss
